# Improve performance of fetchall() and fetchmany()