**Fixed performance issue in fetchall() and fetchmany()**

On a sample data (100000 rows):
**Before**:
```python
%timeit conn.execute(q).fetchall()
6.46 s ± 228 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

**After**:
```python
%timeit conn.execute(q).fetchall()
4.08 s ± 97.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
**Elapsed time decreased by 37%**